Refactor GetProductsList. Cleanup using statements.
