+++ Src/Application/Categories/Queries/GetCategoriesList/CategoriesListVm.cs
@@ -1,6 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
+﻿using System.Collections.Generic;
 
 namespace Northwind.Application.Categories.Queries.GetCategoriesList
 {
