+++ Src/Application/Categories/Queries/GetCategoriesList/GetCategoriesListQuery.cs
@@ -1,6 +1,4 @@
-﻿using System.Collections.Generic;
-using System.Linq;
-using MediatR;
+﻿using MediatR;
 
 namespace Northwind.Application.Categories.Queries.GetCategoriesList
 {
