+++ Src/Application/Categories/Queries/GetCategoriesList/GetCategoriesListQueryHandler.cs
@@ -1,10 +1,10 @@
-﻿using System.Threading;
-using System.Threading.Tasks;
-using AutoMapper;
+﻿using AutoMapper;
 using AutoMapper.QueryableExtensions;
 using MediatR;
 using Microsoft.EntityFrameworkCore;
 using Northwind.Application.Common.Interfaces;
+using System.Threading;
+using System.Threading.Tasks;
 
 namespace Northwind.Application.Categories.Queries.GetCategoriesList
 {
