+++ Src/Application/Customers/Commands/CreateCustomer/CreateCustomerCommand.cs
@@ -1,8 +1,8 @@
-﻿using System.Threading;
-using System.Threading.Tasks;
-using MediatR;
+﻿using MediatR;
 using Northwind.Application.Common.Interfaces;
 using Northwind.Domain.Entities;
+using System.Threading;
+using System.Threading.Tasks;
 
 namespace Northwind.Application.Customers.Commands.CreateCustomer
 {
