+++ Src/Application/Customers/Commands/DeleteCustomer/DeleteCustomerCommandHandler.cs
@@ -1,10 +1,10 @@
-﻿using System.Linq;
-using System.Threading;
-using System.Threading.Tasks;
-using MediatR;
+﻿using MediatR;
 using Northwind.Application.Common.Exceptions;
 using Northwind.Application.Common.Interfaces;
 using Northwind.Domain.Entities;
+using System.Linq;
+using System.Threading;
+using System.Threading.Tasks;
 
 namespace Northwind.Application.Customers.Commands.DeleteCustomer
 {
