--- Src/Application/Products/Queries/GetAllProducts/GetAllProductsQuery.cs
@@ -1,8 +1,0 @@
-﻿using MediatR;
-
-namespace Northwind.Application.Products.Queries.GetAllProducts
-{
-    public class GetAllProductsQuery : IRequest<ProductsListViewModel>
-    {
-    }
-}
