--- Src/Application/Products/Queries/GetAllProducts/ProductsListViewModel.cs
@@ -1,11 +1,0 @@
-﻿using System.Collections.Generic;
-
-namespace Northwind.Application.Products.Queries.GetAllProducts
-{
-    public class ProductsListViewModel
-    {
-        public IEnumerable<ProductDto> Products { get; set; }
-
-        public bool CreateEnabled { get; set; }
-    }
-}
