+++ Src/Application/Products/Queries/GetProductsFile/GetProductsFileQuery.cs
@@ -1,7 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-using MediatR;
+﻿using MediatR;
 
 namespace Northwind.Application.Products.Queries.GetProductsFile
 {
