+++ Src/Application/Products/Queries/GetProductsFile/GetProductsFileQueryHandler.cs
@@ -1,11 +1,11 @@
-﻿using System.Threading;
-using System.Threading.Tasks;
-using AutoMapper;
+﻿using AutoMapper;
 using AutoMapper.QueryableExtensions;
 using MediatR;
 using Microsoft.EntityFrameworkCore;
 using Northwind.Application.Common.Interfaces;
 using Northwind.Common;
+using System.Threading;
+using System.Threading.Tasks;
 
 namespace Northwind.Application.Products.Queries.GetProductsFile
 {
