+++ Src/Application/Products/Queries/GetProductsList/GetProductsListQuery.cs
@@ -1,0 +1,8 @@
+﻿using MediatR;
+
+namespace Northwind.Application.Products.Queries.GetProductsList
+{
+    public class GetProductsListQuery : IRequest<ProductsListVm>
+    {
+    }
+}
