+++ Src/Application/Products/Queries/GetProductsList/ProductsListVm.cs
@@ -1,0 +1,11 @@
+﻿using System.Collections.Generic;
+
+namespace Northwind.Application.Products.Queries.GetProductsList
+{
+    public class ProductsListVm
+    {
+        public IList<ProductDto> Products { get; set; }
+
+        public bool CreateEnabled { get; set; }
+    }
+}
