+++ Src/Application/System/Commands/SeedSampleData/SampleDataSeeder.cs
@@ -1,11 +1,11 @@
 ﻿using Microsoft.EntityFrameworkCore;
+using Northwind.Application.Common.Interfaces;
+using Northwind.Domain.Entities;
 using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
-using Northwind.Application.Common.Interfaces;
-using Northwind.Domain.Entities;
 
 namespace Northwind.Persistence
 {
